+++ client/components/basic/userStatus/UserStatus.js
@@ -1,12 +1,7 @@
 import React from 'react';
 import { Box } from '@rocket.chat/fuselage';
 
-const statusColors = {
-	offline: 'neutral-500',
-	busy: 'danger-500',
-	away: 'warning-500',
-	online: 'success-500',
-};
+import statusColors from '../../../helpers/statusColors';
 
 const UserStatus = React.memo(({ status, ...props }) => <Box size='x12' borderRadius='full' backgroundColor={statusColors[status]} {...props}/>);
