+++ client/helpers/statusColors.js
@@ -1,0 +1,8 @@
+const statusColors = {
+	offline: 'neutral-500',
+	busy: 'danger-500',
+	away: 'warning-500',
+	online: 'success-500',
+};
+
+export default statusColors;
